+++ Realm/ObjectStore/results.cpp
@@ -85,15 +85,21 @@ Results::~Results()
     }
 }
 
-void Results::validate_read() const
+bool Results::is_valid() const
 {
     if (m_realm)
         m_realm->verify_thread();
+
     if (m_table && !m_table->is_attached())
-        throw InvalidatedException();
-    if (m_mode == Mode::TableView && (!m_table_view.is_attached() || m_table_view.depends_on_deleted_object()))
-        throw InvalidatedException();
-    if (m_mode == Mode::LinkView && !m_link_view->is_attached())
+        return false;
+
+    return true;
+}
+
+void Results::validate_read() const
+{
+    // is_valid ensures that we're on the correct thread.
+    if (!is_valid())
         throw InvalidatedException();
 }