+++ Realm/RLMResults.mm
@@ -126,6 +126,10 @@ static inline void RLMResultsValidateInWriteTransaction(__unsafe_unretained RLMR
     ar->_realm->_realm->verify_in_write();
 }
 
+- (BOOL)isInvalidated {
+    return translateErrors([&] { return !_results.is_valid(); });
+}
+
 - (NSUInteger)count {
     return translateErrors([&] { return _results.size(); });
 }